refactor, move middlewares
