+++ hw12_13_14_15_calendar/internal/server/http/server.go
@@ -21,57 +21,9 @@ const userIdKey = "userId"
 
 type BasicHandler func(http.ResponseWriter, *http.Request)
 
-func logMiddleware(h BasicHandler) BasicHandler {
-	return func(w http.ResponseWriter, r *http.Request) {
-		defer func(t time.Time) {
-			log.Println(r.RemoteAddr+" "+r.Method+" "+r.Host+" "+r.UserAgent(), " ", time.Since(t).Milliseconds(), "ms")
-		}(time.Now())
-
-		h(w, r)
-	}
-}
-
 // check requered fields
 // compose event with coerce
 // check fields validity
-
-func dbMiddleware(h BasicHandler, repo repository.BaseRepo) BasicHandler {
-	return func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-		ctx = context.WithValue(ctx, repositoryKey, repo)
-
-		h(w, r.WithContext(ctx))
-	}
-}
-
-func userIdMiddleware(h BasicHandler) BasicHandler {
-	return func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-		userIdStr := r.Header.Get("userid")
-
-		if userIdStr == "" {
-			http.Error(w, "please specify userId in headers", http.StatusUnauthorized)
-			return
-		}
-
-		userId, err := strconv.Atoi(userIdStr)
-
-		if err != nil {
-			http.Error(w, "please check your userId", http.StatusUnauthorized)
-			return
-		}
-
-		ctx = context.WithValue(ctx, userIdKey, userId)
-		h(w, r.WithContext(ctx))
-	}
-}
-
-func applyMiddlewares(h BasicHandler, r repository.BaseRepo) BasicHandler {
-	h1 := dbMiddleware(h, r)
-	h2 := userIdMiddleware(h1)
-
-	return logMiddleware(h2)
-}
 
 func helloHandler(w http.ResponseWriter, req *http.Request) {
 	fmt.Fprintf(w, "hello world\n")
